// halfserious/src/app/interfaces/spaceships.interface.ts
export interface StarshipElement {
  model: string;
  created: string;
  edited: string;
  manufacturer: string;
  cost_in_credits: Array<string>;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  consumables: string;
  passengers: string;
  cargo_capacity: string;
  name: string;
  MGLT: string;
  hyperdrive_rating: string;
  starship_class: string;
  pilots: Array<string>;
  starships: Array<string>;
  url: string;
  films: Array<string>;
}